requirements.txt
# Основные зависимости для AI Telegram Bot
python-telegram-bot==20.7
pandas==2.1.4
numpy>=1.26.0
openpyxl==3.1.2
python-dotenv==1.0.0

# OpenAI для ChatGPT интеграции
openai>=1.0.0

# Дополнительные зависимости
requests>=2.31.0
